Door Alpha & Interior Structure
Added beginning phase of doors (can be opened/closed by pressing E.)

In addition, there is a camera that can be activated in the door if the player crouches and presses E.

Currently there is a bug with the system where if a player uses the door camera, the game thinks they aren't crouching while they are and vice versa causing sprinting to break.

Lastly The interior of the house is mostly completed and textured. Small edits remain, but decorating can begin.